refactor: 주간분석이 존재하지 않을때 예외를 NoSuchElementEx -> WeeklyReportNotFoundEx 로 변경
