+++ src/main/java/bsise/server/error/WeeklyReportNotFoundException.java
@@ -1,0 +1,10 @@
+package bsise.server.error;
+
+import jakarta.persistence.EntityNotFoundException;
+
+public class WeeklyReportNotFoundException extends EntityNotFoundException {
+
+    public WeeklyReportNotFoundException(String message) {
+        super(message);
+    }
+}
